test: Moving `utils.ts` into `tests/utils`
Moved `tests/utils.ts` to `tests/utils/utils.ts` and created `test/utils/index.ts`. Trying to keep the utils within one place.

Removed `setupGlobalAgent` code, #420 means it isn't used anymore.

Updating how we handle conditional testing utility's. we're switching to using reified booleans and just composing conditional boolean expression.

#434
