+++ tests/utils/index.ts
@@ -1,0 +1,2 @@
+export * from './utils';
+export * as exec from './exec';
